interfaz valida falta scrollpanel

+++ src/Chromosome.java
@@ -1,6 +1,6 @@
 
 public interface Chromosome extends Comparable<Chromosome>{
 	public void mutate();
-	public int eval();
+	public double eval();
 	public Chromosome join(Chromosome aCouple);
 }
